Extra Tree Classifier for Feature Selection
Extremely Randomized Trees Classifier(Extra Trees Classifier) is a type of ensemble learning
technique which aggregates the results of multiple de-correlated decision trees collected in
a “forest” to output it’s classification result. In concept, it is very similar to a
Random Forest Classifier and only differs from it in the manner of construction of the decision
trees in the forest.

Each Decision Tree in the Extra Trees Forest is constructed from the original training sample. 
Then, at each test node, Each tree is provided with a random sample of k features from the 
feature-set from which each decision tree must select the best feature to split the data based
on some mathematical criteria (typically the Gini Index). This random sample of features leads
to the creation of multiple de-correlated decision trees.